requirements.txt
google-api-python-client>=2.70.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
twilio>=8.10.0
watchdog>=3.0.0
requests>=2.28.0